Fixing transactions
